requirements.txt
aiofiles==22.1.0
aiosmtplib==2.0.1
anyio==3.6.2
bcrypt==4.0.1
black==22.12.0
blinker==1.5
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.89.1
fastapi-jwt-auth==0.5.0
fastapi-login==1.8.3
fastapi-mail==1.2.4
Flask==2.2.2
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.3
platformdirs==2.6.2
pyasn1==0.4.8
pydantic==1.10.4
PyJWT==1.7.1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
zipp==3.11.0
